Allow user to get first letter of word and catch exceptions for number players and difficulty

+++ src/com/company/Scores.java
@@ -12,7 +12,7 @@ public class Scores{
 
     //TODO: Currently adding a new player -> if name already exists, get that name and add to it.
 
-    public void addScore(String name, int points){
+    public void addScore(String name, double points){
         scores.put(name, points);
     }
